Move IShaderProgram interface to new Velaptor.OpenGL.Shaders namespace & folder
