+++ Testing/VelaptorTests/Graphics/SpriteBatchTests.cs
@@ -17,6 +17,8 @@ namespace VelaptorTests.Graphics
     using Velaptor.Graphics;
     using Velaptor.NativeInterop.OpenGL;
     using Velaptor.OpenGL;
+    using Velaptor.OpenGL.Buffers;
+    using Velaptor.OpenGL.Shaders;
     using Velaptor.Reactables.Core;
     using Velaptor.Reactables.ReactableData;
     using Velaptor.Services;
