+++ Velaptor/Graphics/SpriteBatch.cs
@@ -16,6 +16,8 @@ namespace Velaptor.Graphics
     using Velaptor.Content.Fonts;
     using Velaptor.NativeInterop.OpenGL;
     using Velaptor.OpenGL;
+    using Velaptor.OpenGL.Buffers;
+    using Velaptor.OpenGL.Shaders;
     using Velaptor.Reactables.Core;
     using Velaptor.Reactables.ReactableData;
     using Velaptor.Services;
